pyproject.toml
[project]
name = "o365proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
   "flask<3",
   "requests<3",
   "python-dotenv<2"
]
